% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.weight.R
\name{get.weight}
\alias{get.weight}
\title{Weight function for  a binary alteration matrix of genes and patients.
As described in http://www.ncbi.nlm.nih.gov/pmc/articles/PMC3266044/}
\usage{
get.weight(bin.mutmat)
}
\arguments{
\item{bin.mutmat}{logical matrix, genes as rows, samples as columns}
}
\description{
Weight function for  a binary alteration matrix of genes and patients.
As described in http://www.ncbi.nlm.nih.gov/pmc/articles/PMC3266044/
}
