% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoprint.R
\name{incrementMatrix}
\alias{incrementMatrix}
\title{Function to increment gene matrix}
\usage{
incrementMatrix(M, events, scoring)
}
\arguments{
\item{M}{data frame, colnames=samples, rownames=genes}

\item{events}{data frame. Needs columns sample and gene}

\item{scoring}{number by which to increment}
}
\description{
Function to increment gene matrix
}
