% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/me.test.fisher.R
\name{me.test.fisher}
\alias{me.test.fisher}
\title{Function to calculate a mutual exclusion p-value in a gene set.
It performs a Fisher's test for each gene agains the aggregation
of the rest. After adjusting the p-values, it returns the maximum one.}
\usage{
me.test.fisher(mutmat)
}
\arguments{
\item{mutmat}{matrix, genes as rows, samples as columns}
}
\description{
Function to calculate a mutual exclusion p-value in a gene set.
It performs a Fisher's test for each gene agains the aggregation
of the rest. After adjusting the p-values, it returns the maximum one.
}

