% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me.test.permutateSamples.R
\name{me.test.permutateSamples}
\alias{me.test.permutateSamples}
\title{Function to calculate a mutual exclusion p-value in a gene set.
It keeps the number of alterations in a gene fixed, and permutates
the samples to calculate an empirical p-value. The minimum p-value
it will return is 1/n.}
\usage{
me.test.permutateSamples(mutmat, n = 10000)
}
\arguments{
\item{mutmat}{matrix, genes as rows, samples as columns}

\item{n}{integer specifying the number of permutations}
}
\description{
Function to calculate a mutual exclusion p-value in a gene set.
It keeps the number of alterations in a gene fixed, and permutates
the samples to calculate an empirical p-value. The minimum p-value
it will return is 1/n.
}
