% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{tcga_brca}
\alias{tcga_brca}
\title{Description of the TCGA BRCA.}
\format{A dataframe with 3 columns:
\describe{
  \item{gene}{Gene with the alteration (BRCA1 or BRCA2).}
  \item{sample}{Sample id.}
  \item{alteration}{Alteration detected in that gene (AMP for amplification, MUT for somatic mutation, HOMDEL for homozygous deletion).}
}}
\description{
Dataframe containing the mutational status of BRCA1 and BRCA2 over a subset of patients.
}
