% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoprint.R
\name{oncoprint}
\alias{oncoprint}
\title{function to create an oncoprint plot}
\usage{
oncoprint(M, keys = list(somatic = "MUT", germline = "GERMLINE", amp = "AMP",
  del = "HOMDEL", upreg = "UP", downreg = "DOWN", splicing = "SPLICING"),
  sortGenes = TRUE)
}
\arguments{
\item{M}{mutation matrix}

\item{keys}{list with the following elements: splicing, somatic, 
germline, amp, del, upreg, downreg}

\item{sortGenes}{boolean whether or not to sort the genes, default TRUE}
}
\description{
function to create an oncoprint plot
}
